Pokedb script fix (#12)
* 'PokedexDB.cs' was created in new branch to create a connection to the 'Pokedex Database',

No data has been selected or implemented yet, ONLY connection created. A small region has been created at the bottom of the `PokedexDB` class with potential next steps for the database.

Database file to be added ---

* Pokemon database added to the solution explorer, please check to see if it works for everyone. Otherwise it'll have to get remade :)
The original 'sql' file was corrupted so I imported it as a project.

README.md updated the name of ht '.sql'

* Screenshot of the Pokemon Database was taken and inserted for the README.md,

Database file inserted not yet tested. Image example of the values that can be found in the database/project.

'team-docs folder' added for teammates to import exisiting and new documentation :)

* connectionstring updated -- pls test

* Database updated so that teammates can access it and upload their own onto their own files

* Shorter/Compiled `sql` file containing Pokemon data removed and replaced so that other team members can access and edit it, Connection to be tested through localhost

* Directions for creating your own copy of the database on the `.sql` file for the database.

* Test query created and database connected to `Form1` sucessfully.

* Renamed items

* Rewrote the PokedexDatabase.sql

When this script runs, it deletes the existing PokedexDB and recreates it. Afterwards, it creates the tables and adds data to them.

* Fixed Speed Table

The Speed column was able to become NULL, while similar Pokemon attributes could not be set to NULL. This commit fixes that issue.

---------

Co-authored-by: VacaSama <klc_moore@outlook.com>