Entering select query into the LoadSave Method in the Database Class.
Filling in the blanks will general stuff in the database class.

Select query also added to the GetExistingUser() method, the email and password ARE NOT included in the select query for security purposes. Therefore the User class has been modified to no longer require the email and password but it will be referenced in the NewUser method.

All methods are filled in the PokedexDb class and testing is needed to be done to make sure all of the methods work as intended.