Database table&pokemon tracking kourtnie (#31)
* Only editing in the Dev Notes under the PokedexDB Class to showcase the brainstorming process I have going on for the sql file and future plans for the database class.
~~Referencing other projects.

GetAllCollectedPokemon() method made but not implemented no functionality at the moment. Though process/ comments in method

* Creating CollectedPokemon table and Users Table in the database needs to be reviewed. Referenced from the database,

~~~~
CollectedPokemon class made to represent a pokemon that the user has collected. Should the PokemonId be included in the properties/fields?

Users class to be made ~ nxt commit.

* User class made with properties/ fields.

Still nothing have been executed in the `sql` file for the database, please have teammates review database changes before execution. :)

Not all of the columns in `CollectedPokemon Table` in the Database are being references. FIX BEFORE EXECUTION !!

* spelling error

* Planning out methods on the PokedexDb class. Nxt - try to open the connection, select user, check to see if user exists.

Current return of the List<CollectedPokemon> GetAllCollectedPokemone(...)... returns nothing.

adding required to some of the properties.

Name "Ash" for Pokemon Trainer Name Placeholder in PokedexDb Class.

* Try catch inserted into the PokedexDB, no major functionality yet.

Database still not executed

* Making a save file table in the database `.sql` file. It will be used to retrieve the time and date that the user wanted to save. Frequent saves are always good, gamers, if you know you know.

Formatting the SaveName so that its pretty generic and the user doesn't always have to insert a save name.

`PokedexDatabase.sql`, column and table updates.
Data INSERTED into the CollectedPokemon table--- stats are all the same with the execption of the Level column

~~~Thought Process Added to the Database~~~

* Database finally executed. Now to use the database to insert or pull everything

Properties added to the SaveFile class, deleted the extra method in the PokedexDb.

updated sample information in the database so that we can do tests with users such as Ash Ketchum. the data should now be viewable.

* Specifiying and Structuring new methods and thought processes of each assigned class.

Enhance PokedexDB with user and save file management

- Updated constructor in `PokedexDB` to clarify database connection testing.
- Added `NewUser` method for creating new users with insert query comments.
- Introduced `GetExistingUser` method to retrieve user data by ID.
- Added `NewSave` method for saving user progress with insert query comments.
- Implemented `LoadSave` method to retrieve saved files by ID.
- Updated `SaveFile` class with comments describing its purpose.

* Entering select query into the LoadSave Method in the Database Class.

Filling in the blanks will general stuff in the database class.

Select query also added to the GetExistingUser() method, the email and password ARE NOT included in the select query for security purposes. Therefore the User class has been modified to no longer require the email and password but it will be referenced in the NewUser method.

All methods are filled in the PokedexDb class and testing is needed to be done to make sure all of the methods work as intended.

* dev notes deleted in pokedexdb.cs

* summaries updated

* Update PokedexDB.cs

missing curly braces again